+++ mitra_activitypub/src/builders/mod.rs
@@ -1,5 +1,6 @@
 pub mod accept_follow;
 pub mod accept_offer;
+pub mod add_context_activity;
 pub mod add_note;
 pub mod add_person;
 mod agreement;