+++ mitra_activitypub/src/identifiers.rs
@@ -105,6 +105,10 @@ pub fn local_tag_collection(instance_url: &str, tag_name: &str) -> String {
     format!("{}/collections/tags/{}", instance_url, url_encode(tag_name))
 }
 
+pub fn local_context_collection(instance_url: &str, conversation_id: Uuid) -> String {
+    format!("{}/collections/conversations/{}/context", instance_url, conversation_id)
+}
+
 pub fn local_activity_id(
     instance_url: &str,
     activity_type: &str,